Move test scripts
