+++ test_scripts/API/TTSChunks/007_Speak.lua
@@ -9,33 +9,25 @@
 -- Description:
 -- In case:
 -- 1) HMI provides ‘FILE’ item in ‘speechCapabilities’ parameter of ‘TTS.GetCapabilities’ response
--- 2) New app registers and send AlertManeuver with ‘FILE’ item in ‘ttsChunks’ parameter
+-- 2) New app registers and send Speak with ‘FILE’ item in ‘ttsChunks’ parameter
 -- SDL does:
 -- 1) Send TTS.Speak request to HMI with ‘FILE’ item in ‘ttsChunks’ parameter
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
-local common = require('test_scripts/TTSChunks/common')
+local common = require('test_scripts/API/TTSChunks/common')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Functions ]]
-local function pTUpdateFunc(pTbl)
-  table.insert(pTbl.policy_table.app_policies[common.getConfigAppParams().appID].groups, "Navigation-1")
-end
-
-local function sendAlertManeuver()
+local function sendSpeak()
   local params = {
     ttsChunks = {
       { type = common.type, text = "pathToFile" }
     }
   }
-  local corId = common.getMobileSession():SendRPC("AlertManeuver", params)
-  common.getHMIConnection():ExpectRequest("Navigation.AlertManeuver")
-  :Do(function(_, data)
-      common.getHMIConnection():SendResponse(data.id, data.method, "SUCCESS", {})
-    end)
+  local corId = common.getMobileSession():SendRPC("Speak", params)
   common.getHMIConnection():ExpectRequest("TTS.Speak", { ttsChunks = params.ttsChunks })
   :Do(function(_, data)
       common.getHMIConnection():SendResponse(data.id, data.method, "SUCCESS", {})
@@ -48,11 +40,10 @@ runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
 runner.Step("Start SDL, init HMI, connect Mobile", common.start)
 runner.Step("Register App", common.registerApp)
-runner.Step("PTU", common.policyTableUpdate, { pTUpdateFunc })
 runner.Step("Activate App", common.activateApp)
 
 runner.Title("Test")
-runner.Step("Send AlertManeuver", sendAlertManeuver)
+runner.Step("Send Speak", sendSpeak)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
