+++ test_sets/audio_file_playback_tts_chunks.txt
@@ -1,9 +1,9 @@
-./test_scripts/TTSChunks/001_RegisterAppInterface.lua
-./test_scripts/TTSChunks/002_SetGlobalProperties.lua
-./test_scripts/TTSChunks/003_OnAppRegistered.lua
-./test_scripts/TTSChunks/004_Alert.lua
-./test_scripts/TTSChunks/005_ChangeRegistration.lua
-./test_scripts/TTSChunks/006_PerformInteraction.lua
-./test_scripts/TTSChunks/007_Speak.lua
-./test_scripts/TTSChunks/008_PerformAudioPassThru.lua
-./test_scripts/TTSChunks/009_AlertManeuver.lua
+./test_scripts/API/TTSChunks/001_RegisterAppInterface.lua
+./test_scripts/API/TTSChunks/002_SetGlobalProperties.lua
+./test_scripts/API/TTSChunks/003_OnAppRegistered.lua
+./test_scripts/API/TTSChunks/004_Alert.lua
+./test_scripts/API/TTSChunks/005_ChangeRegistration.lua
+./test_scripts/API/TTSChunks/006_PerformInteraction.lua
+./test_scripts/API/TTSChunks/007_Speak.lua
+./test_scripts/API/TTSChunks/008_PerformAudioPassThru.lua
+./test_scripts/API/TTSChunks/009_AlertManeuver.lua
